enable full BF16 mode

+++ examples/moviegen/mg/models/text_encoders/text_projector.py
@@ -22,13 +22,13 @@ class TextProjector(nn.Cell):
         super().__init__()
         # split layers for easier exclusion from weight decay
         self.ul2_linear = nn.Dense(ul2_in_features, out_features, has_bias=False, dtype=dtype)
-        self.ul2_layernorm = layer_norm((out_features,), eps=norm_eps)
+        self.ul2_layernorm = layer_norm((out_features,), eps=norm_eps, dtype=dtype)
 
         self.metaclip_linear = nn.Dense(metaclip_in_features, out_features, has_bias=False, dtype=dtype)
-        self.metaclip_layernorm = layer_norm((out_features,), eps=norm_eps)
+        self.metaclip_layernorm = layer_norm((out_features,), eps=norm_eps, dtype=dtype)
 
         self.byt5_linear = nn.Dense(byt5_in_features, out_features, has_bias=False, dtype=dtype)
-        self.byt5_layernorm = layer_norm((out_features,), eps=norm_eps)
+        self.byt5_layernorm = layer_norm((out_features,), eps=norm_eps, dtype=dtype)
 
         self.initializer_range = initializer_range
